Seemingly refactored. Hopeful that papiv2alpha1 still works
